+++ Makefile
@@ -13,25 +13,26 @@
 DIRS=$(shell go list -f {{.Dir}} ./...)
 DEPEND=\
 	bitbucket.org/pkg/inflect \
-	github.com/PuerkitoBio/purell \
 	github.com/asaskevich/govalidator \
+	github.com/dgrijalva/jwt-go \
+	github.com/dimfeld/httptreemux \
 	github.com/go-swagger/go-swagger \
 	github.com/goadesign/encoding \
+	github.com/goadesign/goa-cellar \
 	github.com/goadesign/goa.design/tools/godoc2md \
 	github.com/goadesign/goa.design/tools/mdc \
-	github.com/goadesign/goa-cellar \
 	github.com/goadesign/middleware \
 	github.com/golang/lint/golint \
-	github.com/dimfeld/httptreemux \
 	github.com/manveru/faker \
 	github.com/on99/gocyclo \
 	github.com/onsi/ginkgo \
 	github.com/onsi/ginkgo/ginkgo \
 	github.com/onsi/gomega \
+	github.com/PuerkitoBio/purell \
 	github.com/spf13/hugo \
 	github.com/zach-klippenstein/goregen \
 	golang.org/x/tools/cmd/cover \
- 	golang.org/x/tools/cmd/goimports \
+	golang.org/x/tools/cmd/goimports \
 	gopkg.in/alecthomas/kingpin.v2 \
 	gopkg.in/tylerb/graceful.v1 \
 	gopkg.in/yaml.v2
