--- middleware/.travis.yml
@@ -1,34 +1,0 @@
-language: go
-go:
-- 1.5.3
-- 1.6
-sudo: false
-install:
-- go get -v ./...
-- export PATH=${PATH}:${HOME}/gopath/bin
-- curl https://bitbucket.org/birkenfeld/pygments-main/get/2.1.tar.gz -L -o 2.1.tar.gz
-- mkdir -p ${HOME}/bin
-- tar -xf 2.1.tar.gz
-- mv birkenfeld-pygments-main-34530db252d3/* ${HOME}/bin
-- rm -rf birkenfeld-pygments-main-34530db252d3
-script:
-- export PATH=${PATH}:${HOME}/bin
-- make
-- make docs
-notifications:
-  slack:
-    secure: tcNpappojuA4O9obbQBX4uyph8J00UyK57yjdYzzVW4WDAFx8KBuaLV7GJoL3c3+nUSwOrp2XzF65JZdQIWf4sjMvIc5wxhPUWXWHY/zIHnlrA3DrOBmr+fEemzUjkPiexfu5C2jwqPNrUyNQchxIGJ+kN3xvHPgXAVdLWQaumO9tVA56VPQlxNfDSPpbuwrEURlWpRBsOYJ8MKD9ZRsgd3FLibc1S8vGb/kIYauQRqqMUg7eBiNigVMFZzcGhNHuHmRiZtHch4HQRNT/SBHga8/6kFfT25UuGDI1JFZVT46nO4r3fGQ5ZnC7f016JjCvVBHJcfw4NDJeSFGTECiJ0WFT6qs+m5dVVKuWZsjJU/92hRMshaK73OEvjljAy9hwapNQj2DMgx38E6zVfhoJn6q+yHYfuXf5Y3x6XscGXlhKpFhgY6o4OvbumggoCBaPPl9avgImifB016aSg8NCPoWODwA0HP+nwEzSbeczEj03gvk27c+KixAKse6vLmdMhtFsd2cloJdjDngrXsDote/vuAF9WWy8KcFDRlNkc+7RIEzNvJYDNS5+yxZAA7gr6eLo9H67lcsJPK9ljz5ZodGCZxmcXYtH/yC1dO38AKyXLZnwJkT8wcbqTJmpdCw7ME+dTY8xHiyCLwmyxI9cOlAN5HTGSEujxYWS7fCnT8=
-deploy:
-  provider: gcs
-  access_key_id: GOOGDIIIVPY7O6DG3PSZ
-  secret_access_key:
-    secure: iuK0ot00MjO7sKFJxkESKhEwIyRV4xD+Xcojaf4KtCXascRdH8R9Z2S16q57hIuDK2jvZhgkzm/5Q+y4/rbqK1Wzis6HnQmkMtoRpiCFQA1bT7Vn+9OzDpGRsnyeS68uNmn/gHhmfySoCL1/W+W88SDePgjrt1FtkA0jPmb5dKcRzGTF01j2lP5Coll0mduw4n7R3DMb04Pzqk4lcqZT0BLFxr7P4GR+xQsFqXXv/gZMFjtaVCD1VgTj0NXdJKuSHJP4+e12w/eGeR9+dDj7IN8yZhM9xh3tkSeBtZCqh9McG4H1goctUs89NIDkgT0cToXB+XP9eXBF//sPDfA+dj+gAhM3qUWyNp1pSWQQ7al+AilYYSbIBYWsbfElcsxrzhjs4EHYjTl4+pTA2WaGxO2JA22/kz9r645CIhumaeuNQ1eGBEj+1XQQrCiUc1WwAFmknBtTV3mLZ6lu2JzZf/mldkItDz4RpEjS3NtMqoqQellOv5yaCnO+iqZq25Wy3/1n+X50LhyvwfWgCW/Rq/aEwPAsMG6qVuaAJdf4y5hn6X3PdzyTm/0YWhnj2iUW9B7lMVLERph3HNgPlMMBUYJWAAjwXn3IgIRxLHrRXILkaatvgKAuwqX7DSyaXSl2aBPlIKyg+iN7TlFJE5DPoWTpMOGa5zEjNa6BDKClTSo=
-  bucket: goa.design
-  local-dir: public/reference
-  upload-dir: reference
-  skip-cleanup: true
-  acl: public-read
-  cache-control: max-age=300
-  on:
-    repo: goadesign/middleware
-    go: '1.5.3'
